requirements.txt
# basic simulation requirements
omegaconf
pygame
pymunk
opencv-python
imantics
pillow
kornia
av
# seine requirements
numpy==1.26.4
matplotlib==3.9.2
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
decord==0.6.0
diffusers==0.15.0
imageio==2.29.0
transformers==4.29.2
xformers==0.0.20
einops
tensorboard==2.15.1
timm==0.9.10
rotary-embedding-torch==0.3.5
natsort==8.4.0



